move _scale_lr to single device strategy
